+++ backend/src/main/java/moaon/backend/project/repository/CustomizedProjectRepositoryImpl.java
@@ -5,56 +5,37 @@ import static moaon.backend.project.domain.QProjectCategory.projectCategory;
 import static moaon.backend.techStack.domain.QTechStack.techStack;
 
 import com.querydsl.core.BooleanBuilder;
-import com.querydsl.core.types.OrderSpecifier;
-import com.querydsl.core.types.dsl.BooleanExpression;
-import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import moaon.backend.global.cursor.ProjectCursor;
-import moaon.backend.global.domain.SearchKeyword;
 import moaon.backend.project.domain.Project;
-import moaon.backend.project.domain.ProjectSortType;
 import moaon.backend.project.dto.ProjectQueryCondition;
+import moaon.backend.project.repository.querymodifier.CategoriesModifier;
+import moaon.backend.project.repository.querymodifier.CursorModifier;
+import moaon.backend.project.repository.querymodifier.SearchKeywordModifier;
+import moaon.backend.project.repository.querymodifier.SortByModifier;
+import moaon.backend.project.repository.querymodifier.TechStacksModifier;
 import org.springframework.stereotype.Repository;
-import org.springframework.util.CollectionUtils;
 
 @Repository
 @RequiredArgsConstructor
 public class CustomizedProjectRepositoryImpl implements CustomizedProjectRepository {
 
-    private static final double MINIMUM_MATCH_SCORE = 0.0;
-    private static final String BLANK = " ";
     private static final int FETCH_EXTRA_FOR_HAS_NEXT = 1;
 
     private final JPAQueryFactory jpaQueryFactory;
 
     @Override
     public List<Project> findWithSearchConditions(ProjectQueryCondition condition) {
-        ProjectCursor<?> cursor = condition.cursor();
-
         JPAQuery<Project> query = jpaQueryFactory.selectFrom(project)
                 .distinct()
                 .leftJoin(project.categories, projectCategory)
                 .leftJoin(project.techStacks, techStack);
 
-        BooleanBuilder whereBuilder = new BooleanBuilder();
-
-        applyWhereAndHaving(whereBuilder, condition, query);
-
-        if (cursor != null) {
-            cursor.applyCursor(condition, whereBuilder);
-        }
-
-        if (whereBuilder.hasValue()) {
-            query.where(whereBuilder);
-        }
-
         return query.groupBy(project.id)
-                .orderBy(toOrderBy(condition.projectSortType()))
+                .where(constructWhereClause(condition, query))
+                .orderBy(new SortByModifier().modify(condition))
                 .limit(condition.limit() + FETCH_EXTRA_FOR_HAS_NEXT)
                 .fetch();
     }
@@ -66,67 +47,23 @@ public class CustomizedProjectRepositoryImpl implements CustomizedProjectReposit
                 .leftJoin(project.categories, projectCategory)
                 .leftJoin(project.techStacks, techStack);
 
-        BooleanBuilder whereBuilder = new BooleanBuilder();
-
-        applyWhereAndHaving(whereBuilder, condition, query);
-
-        if (whereBuilder.hasValue()) {
-            query.where(whereBuilder);
-        }
-
         return query.groupBy(project.id)
+                .where(constructWhereClause(condition, query))
                 .fetch()
                 .size();
     }
 
-    private void applyWhereAndHaving(
-            BooleanBuilder whereBuilder,
+    private BooleanBuilder constructWhereClause(
             ProjectQueryCondition queryCondition,
             JPAQuery<?> query
     ) {
-        List<String> techStackNames = queryCondition.techStackNames();
-        List<String> categoryNames = queryCondition.categoryNames();
+        BooleanBuilder whereBuilder = new BooleanBuilder();
 
-        if (!CollectionUtils.isEmpty(techStackNames)) {
-            whereBuilder.and(techStack.name.in(techStackNames));
-            query.having(techStack.name.countDistinct().eq((long) techStackNames.size()));
-        }
+        new TechStacksModifier(query, whereBuilder).modify(queryCondition);
+        new CategoriesModifier(query, whereBuilder).modify(queryCondition);
+        new SearchKeywordModifier(whereBuilder).modify(queryCondition);
+        new CursorModifier(whereBuilder).modify(queryCondition);
 
-        if (!CollectionUtils.isEmpty(categoryNames)) {
-            whereBuilder.and(projectCategory.name.in(categoryNames));
-            query.having(projectCategory.name.countDistinct().eq((long) categoryNames.size()));
-        }
-
-        SearchKeyword searchKeyword = queryCondition.search();
-        if (searchKeyword.hasValue()) {
-            whereBuilder.and(satisfiesMatchScore(searchKeyword));
-        }
-    }
-
-    private BooleanExpression satisfiesMatchScore(SearchKeyword searchKeyword) {
-        return Expressions.numberTemplate(
-                Double.class,
-                ProjectFullTextSearchHQLFunction.EXPRESSION_TEMPLATE,
-                formatSearchKeyword(searchKeyword)
-        ).gt(MINIMUM_MATCH_SCORE);
-    }
-
-    private String formatSearchKeyword(SearchKeyword searchKeyword) {
-        String search = searchKeyword.replaceSpecialCharacters(BLANK);
-        return Arrays.stream(search.split(BLANK))
-                .map(keyword -> String.format("+%s*", keyword.toLowerCase()))
-                .collect(Collectors.joining(BLANK));
-    }
-
-    private OrderSpecifier<?>[] toOrderBy(ProjectSortType sortBy) {
-        if (sortBy == ProjectSortType.CREATED_AT) {
-            return new OrderSpecifier<?>[]{project.createdAt.desc(), project.id.desc()};
-        }
-
-        if (sortBy == ProjectSortType.VIEWS) {
-            return new OrderSpecifier<?>[]{project.views.desc(), project.id.desc()};
-        }
-
-        return new OrderSpecifier<?>[]{project.lovedMembers.size().desc(), project.id.desc()};
+        return whereBuilder;
     }
 }
