+++ backend/src/main/java/moaon/backend/project/repository/QueryModifier.java
@@ -1,0 +1,8 @@
+package moaon.backend.project.repository;
+
+import moaon.backend.project.dto.ProjectQueryCondition;
+
+public interface QueryModifier<T> {
+
+    T modify(ProjectQueryCondition condition);
+}
