+++ backend/src/main/java/moaon/backend/project/repository/querymodifier/CategoriesModifier.java
@@ -1,0 +1,29 @@
+package moaon.backend.project.repository.querymodifier;
+
+import static moaon.backend.project.domain.QProjectCategory.projectCategory;
+
+import com.querydsl.core.BooleanBuilder;
+import com.querydsl.jpa.impl.JPAQuery;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import moaon.backend.project.dto.ProjectQueryCondition;
+import moaon.backend.project.repository.QueryModifier;
+import org.springframework.util.CollectionUtils;
+
+@RequiredArgsConstructor
+public final class CategoriesModifier implements QueryModifier<Void> {
+
+    private final JPAQuery<?> query;
+    private final BooleanBuilder whereBuilder;
+
+    @Override
+    public Void modify(ProjectQueryCondition condition) {
+        List<String> categoryNames = condition.categoryNames();
+        if (!CollectionUtils.isEmpty(categoryNames)) {
+            whereBuilder.and(projectCategory.name.in(categoryNames));
+            query.having(projectCategory.name.countDistinct().eq((long) categoryNames.size()));
+        }
+
+        return null;
+    }
+}
