+++ backend/src/main/resources/META-INF/services/org.hibernate.boot.model.FunctionContributor
@@ -1,2 +1,2 @@
 moaon.backend.article.repository.ArticleFullTextSearchHQLFunction
-moaon.backend.project.repository.ProjectFullTextSearchHQLFunction
+moaon.backend.project.repository.querymodifier.SearchKeywordModifier$ProjectFullTextSearchHQLFunction
